+++ release_notes.md
@@ -1,3 +1,5 @@
-* https://github.com/crispinb/cvcap/issues/5
+## release ??
+* Capture text from clipboard
+  https://github.com/crispinb/cvcap/issues/5
 
 ## release-v0.1.0-alpha